refactor(tls): split src/index.ts into per-concern modules (971 → 129 + 8)
Follows the same pattern as the @gjsify/webgl webgl-context-base.ts split
(STATUS.md Workstream D) + the just-landed @gjsify/process split. The
971-line `src/index.ts` becomes a thin 129-line composer on top of 8
focused modules.

**Layout:**

  src/internal/pem.ts         — PEM helpers + Gio.TlsCertificate builders
                                (pemToString, splitPemBlocks,
                                buildGioCertificate, buildCaCertificates,
                                pemToDer)
  src/internal/cert-utils.ts  — Peer-cert extraction (Gio → Node shape):
                                parseDistinguishedName, formatCertDate,
                                formatAltNames, fingerprintFromBytes,
                                tlsCertToPeerCert + PeerCertificate /
                                CertSubject types
  src/internal/hostname.ts    — RFC 6125 §6.4.3 hostname matching
                                (unfqdn, splitHost, isPrintableAscii,
                                checkHostMatch) + checkServerIdentity +
                                CertAltNameError
  src/internal/sni-parser.ts  — (unchanged) ClientHello server_name parser
  src/secure-context.ts       — SecureContext / SecureContextOptions +
                                createSecureContext (factory)
  src/tls-socket.ts           — TLSSocket class + getters
                                (getPeerCertificate, getProtocol,
                                getCipher, getAlpnProtocol) +
                                TlsConnectOptions + SocketInternals
                                (shared with connect.ts + tls-server.ts)
  src/connect.ts              — tls.connect (TLS client): TCP connect →
                                Gio.TlsClientConnection upgrade + mTLS +
                                ALPN + custom-CA validation +
                                customCheckServerIdentity hook
  src/tls-server.ts           — TLSServer class + createServer overloads
                                + SNI peek-and-parse (the BufferedInput
                                Stream + SimpleIOStream flow from #223)
                                + SNICallback type + TlsServerOptions
  src/ocsp.ts                 — (unchanged) parseOcspResponse re-exports
                                from @gjsify/tls-native
  src/index.ts                — composer: header (Node↔Gio mapping table)
                                + DEFAULT_* constants + getCiphers +
                                rootCertificates + re-exports + default
                                tlsExports object

**Module count:** 1 → 9 source files. **Total LoC:** ~971 → 1408
(modest overhead from imports + per-file headers, normal for a split).

**`as unknown as` count: 11 → 11.** All remaining casts are inherent
to crossing the @gjsify/net `SocketInternals` boundary (no way to type
that without making the private fields public on `node:net`). The
split makes WHERE each cast happens more obvious, even if it doesn't
reduce the absolute count.

**`@gjsify/utils` import surface narrowed** — `createNodeError` /
`deferEmit` now only imported by `tls-server.ts` (the one place
that emits `'error'` / `'tlsClientError'` events). `connect.ts` is
@gjsify/utils-free.

**Tests:** 273/273 Node + 180/180 GJS (unchanged — pure structural
refactor, no behavioural change).
